Added some functionality to PhysTimer and PhysGraph. Slightly modified PhysAxis to address user Issue # 6. The individual issue is solved, but this function may still have other things that may need to be addressed. Did not add any unit tests but they all the old ones are passed successfully.
PhysTimer now has an additional argument fontsize, which sets the size of the text. Just call it using PhysTimer(1,2,fontsize = 25) when constructing a PhysTimer object to create a timer object at (x,y) = (1,2) which displays the text in 25 point font. PhysTimer(1,2,25) is equivalent. If fontsize is not specified, the timer will be displayed in size 13 font. Feel free to adjust this; it is just what the default value is for in visual.label

PhysGraph now has additional arguments for horizontal and vertical axes labels and for a window title. Call PhysGraph(2,title = 'Speed vs. Time', xlabel = 't', ylabel = 'v') creates a PhysGraph for two data sets, sets the window title to 'Speed vs. Time', and generates labels on the horizontal and vertical axes that say 't' and 'v', respectively. Alternatively, use PhysGraph(2,'Speed vs. Time','t','x'). Might want to modify code so variable names are hlabel and vlabel. Not sure what is better pedagogically: Use "horizontal and vertical" axes so students know which label is which and don't automatically plot x (if it is position or something) along the horizontal axis or use "x and y" axes to make the connection back to their math classes. I leave it up to you PER types....

Finally, I modified PhysAxis to be centered right below the base object by default, regardless of where that object may be. It works fine for objects that are long in the x direction and thin in the y direction, so it addresses J.Burke's particular example in Issue # 6. This class still fails miserably by default if you use it on vertically extended objects (say box(size(3,4,1)). The labels end up in the middle of nowhere. It looks extra weird if you are trying to make a vertical axis. Naturally, all this can be fixed by tweaking other parameters by hand, but I guess that kind of betrays the "it-just-works" idea behind PhysUtil. BTW, if this mod somehow messes up the overall idea behind PhysAxis and you want to revert to the previous version of PhysAxis, just remove the "+ self.obj.pos" at the end of Line 320.

git-svn-id: https://python-physutil.googlecode.com/svn/trunk@40 2ebeee70-0fbd-7a0e-70c4-9101542a6452